[ci] Re-run failed tests on failure (#12055)
* [ci] Re-run failed tests on failure

This uses [pytest-rerunfailures](https://github.com/pytest-dev/pytest-rerunfailures) to retry failed tests. This should help alleviate flakiness for issues like segfaults in ethosu tests or flaky numerics. This obviously isn't as good as fixing the tests themselves, but it's an easy way to fix the most pressing issue (the cost of developer time to wait for, check, and re-run CI) while keeping the signal from the times the tests do fail consistently.

* Remove changes to pytest

Co-authored-by: driazati <driazati@users.noreply.github.com>