[ci] Fix build android rpc failure (#12216)
android_rpc build problem: https://github.com/apache/tvm/issues/12191

The problem with the build appeared due to the fact that the `ANDROID_NDK_HOME` environment variable was removed in the current version of github actions. 
https://github.com/actions/virtual-environments/blob/main/images/linux/Ubuntu2004-Readme.md

But this variable is used here:

https://github.com/apache/tvm/blob/ee319d9d23c80091da9c4fb764b1e6d49d462714/.github/workflows/main.yml#L122-L127

Now only `ANDROID_NDK_LATEST_HOME` is available for ndk.