[AutoTVM][Fix] Fix wrong axis names of data_vec (#12303)
This PR is trying to fix the wrong axis names of data_vec. As the data_vec is nchwc format, the axis names should be batch, ic_chunk, ih, iw, ic_block, but not batch, ic_chunk, ih, ic_block, iw.
Although the following code does not use these last two axises, so it does not cause some bugs for now. But I think we should fix this.