Fix for import requests and import caffe failures (#12813)
Recently virtual environments were introduced in the
docker images which was a great contribution to
localize errors: https://github.com/apache/tvm/pull/12663. In this fix, link to the caffe is
created inside this virtual env instead of adding it
to the system path of python. This fix also removes
importing request package where not needed.

Fixes #12663