[TIR][Bugfix] Correct handling of buffer argument when scheduling (#12816)
Follow-up from https://github.com/apache/tvm/pull/11269, which allowed
schedule arguments of the buffer to be transformed to be specified as
a string, or as a `tir::Buffer`.  The string handling worked
correctly, but the `tir::Buffer` object was handled incorrectly.  This
commit corrects the handling of `tir::Buffer` arguments when
scheduling, and adds a unit test to validate this behavior.