[TVMScript] Infer T.match_buffer parameters for region (#12890)
* [TVMScript] Infer T.match_buffer parameters for region

When using `T.match_buffer` to define a view into another buffer,
default shape and dtype parameters can be inferred.

* Updated unit test for new behavior

The test intentionally triggers a failed match based on mismatched
`elem_offset`.  Therefore, the test now needs to explicitly pass an
`elem_offset` to trigger the failure, as this now defaults to having a
`Var` for `match_buffer` calls that represent views.