[LLVM] Emit fp16/fp32 builtins directly into target module (#12877)
For conversions between `_Float16` and `float`, LLVM uses runtime functions
`__extendhfsf2` and `__truncsfhf2`.  On X86 up until version 14, LLVM used
`uint16_t` for representing `_Float16`. Starting with LLVM 15, half-
precision values can be passed in XMM registers (i.e. as floating-point).
This happens when the compilation target has SSE2 enabled (either directly,
or by enabling a feature that implies SSE2).
Because the names of the conversion functions remain unchanged, it is
impossible for TVM to provide them in the runtime, and have them work in
both cases. To solve this issue, emit these functions directly into the
target module after detecting whether or not to use floating-point ABI.
To allow the linker to remove potential duplicates (or if they are unused),
they are weak and reside in a separate section.