[ETHOSN] Support conversion of add/mul to requantize where possible (#12887)
* [ETHOSN] Support conversion of add/mul to requantize where possible

Add/mul operations that correspond to identity operations can be
converted to a simple reinterpret quantize operation. This conversion
takes place in the convert equivalents pass similar to the depthwise
counter-part.

In addtion, an issue was noticed that would cause unsupported
operations to raise an error rather than not being offloaded. This has
been fixed by allowing the conversion to return Null when the conversion
is not supported.