[TIR] Use buffer's dtype when converting pad_value to TIR (#12925)
[TIR] Use buffer's dtype when converting pad_value to TIR

Previously, a python int/float would be converted to a TIR
int32/float32 in `tvm.runtime.convert`.  These types may not match the
buffer's type, even though it is the closest Python representation of
the buffer's type.  Therefore, the `pad_value` should be
preferentially converted to the matching type of the buffer.