[TIR][Arith] Implement kApplyConstraintsToBooleanBranches extension (#13129)
When simplifying a branch of a boolean AND or a boolean OR, the
other branch may be assumed not to dominate the result of the
operator.  For example, when simplifying `(A and B)`, `A` may be
simplified on the assumption that `B` is true.  Similarly, when
simplifying `(A or B)`, `A` may be simplified on the assumption
that `B` is false.

Prior to this commit, these constraints were not used for
simplifications.  This commit introduced an optional extension,
`kApplyConstraintsToBooleanBranches`, which exposes these
constraints for simplification.  This isn't enabled by default,
as some cases require a second visit to each branch of a boolean
operator.  (e.g. Simplifying the LHS of an operator, using a
constraint provided by the RHS, where the sub-analyzer that uses
the constraint expects it to already be simplified.)