[microTVM][Zephyr] Add 'serial_number' option (#13377)
This commit adds a new option 'serial_number' to allow specifying the   
serial device number in Zephyr, used to flash and to communicate with   
target boards attached to specific serial ports, and adjusts Zephyr      
tests to run with this new option. This is particularly useful in an    
environment where multiple boards are connected to the host.            
                                                                        
It also removes 'west_cmd' in test arguments, as it is not used, and    
refactors Zephyr server API to set the required/optional project options
at the beginning of each API call.